・チェスト類が買えないように ・棒の工費を安く ・石の強化コンポーネントの素材を安く

+++ src/main/java/net/riblab/tradecore/craft/TCCraftingRecipe.java
@@ -29,7 +29,7 @@ public class TCCraftingRecipe {
     /**
      * レシピを実行するための費用
      */
-    private final int fee;
+    private final double fee;
 
     /**
      * レシピのカテゴリ
